Fix warnings in core layer

+++ Source/core/Types.cpp
@@ -63,11 +63,11 @@ std::string GuidToString(GUID guid) {
 }
 
 void SetDebugObjectName(IDXGIObject* obj, std::string const& name) {
-	ENSURE_HR(obj->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str()));
+	ENSURE_HR(obj->SetPrivateData(WKPDID_D3DDebugObjectName, UINT(name.size()), name.c_str()));
 }
 
 void SetDebugObjectName(ID3D11DeviceChild* res, std::string const& name) {
-	ENSURE_HR(res->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.c_str()));
+	ENSURE_HR(res->SetPrivateData(WKPDID_D3DDebugObjectName, UINT(name.size()), name.c_str()));
 }
 
 rttr::type const& helpers::get_type_by_id(u64 id) {
